feat: fix shiftdate service function input ISOstring instead of year-month-day, change wh_manager menu name, fix timesheetcheck logic, fix simple fingerprint delete

+++ server/src/mqtt/mqtt.controller.ts
@@ -47,7 +47,7 @@ export class MqttController {
 
                     const hasPendingDeletion = await this.mqttService.handleEsp8266DeletionConfirmation(data.employeeID, true);
 
-                    if (!hasPendingDeletion) {
+                    if (!hasPendingDeletion && data.employeeID > 0) {
                         await this.mqttService.deleteEmployeeFingerprint(data.employeeID);
                     }
                 } else {
